requirements.txt
fastapi
uvicorn
python-docx
httpx
python-multipart
